Separation of jvm-specific code from CodecTest
